feat(exports): move export to graphile worker (#738)
