+++ src/server/worker.ts
@@ -17,6 +17,8 @@ import {
   syncCampaignContactToVAN,
   updateVanSyncStatuses
 } from "./tasks/sync-campaign-contact-to-van";
+import exportCampaign from "./tasks/export-campaign";
+import { exportForVan } from "./tasks/export-for-van";
 
 const logFactory: LogFunctionFactory = scope => (level, message, meta) =>
   logger.log({ level, message, ...meta, ...scope });
@@ -55,6 +57,8 @@ export const getWorker = async (attempt = 0): Promise<PgComposeWorker> => {
   m.taskList!["van-get-result-codes"] = fetchVANResultCodes;
   m.taskList!["van-sync-campaign-contact"] = syncCampaignContactToVAN;
   m.taskList!["update-van-sync-statuses"] = updateVanSyncStatuses;
+  m.taskList!["export-campaign"] = exportCampaign;
+  m.taskList!["export-campaign-for-van"] = exportForVan;
 
   m.cronJobs!.push({
     name: "release-stale-replies",
